added route protextion

+++ src/data-access/room.ts
@@ -35,5 +35,14 @@ export async function getRoom(roomId: string) {
 }
 
 export async function deleteRoom(roomId: string) {
+  const session = await auth();
+  const userId = session?.user?.id;
+
+  if (!userId) {
+    throw new Error("userId is not found");
+  }
+  if (!session) {
+    throw new Error("You must be logged in to create this room");
+  }
   await database.delete(room).where(eq(room.id, roomId));
 }
